Cleanup
Remove old weather undergrand code as that feed was turned off and rename new implementation to be what the old one was called.
